chore: move PinnedFQAButton to components folder
